requirements.txt
# Add your project dependencies here
kiteconnect>=4.0.0
python-dotenv>=1.0.1
pyodbc>=5.1.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0